@@ -1,4 +1,4 @@
 import {expectType} from 'tsd';
-import pImmediate = require('.');
+import pImmediate from './index.js';
 
 expectType<Promise<void>>(pImmediate());
